requirements.txt
flask
flask-cors
requests
ollama
pdfplumber
python-docx